@@ -7,15 +7,15 @@
     }
   ],
   "dependencies": {
-    "@angular/animations": "18.2.3",
+    "@angular/animations": "19.0.3",
     "@angular/cdk": "18.2.3",
-    "@angular/common": "18.2.3",
-    "@angular/compiler": "18.2.3",
-    "@angular/core": "18.2.3",
-    "@angular/forms": "18.2.3",
+    "@angular/common": "19.0.3",
+    "@angular/compiler": "19.0.3",
+    "@angular/core": "19.0.3",
+    "@angular/forms": "19.0.3",
     "@angular/material": "18.2.3",
-    "@angular/platform-browser": "18.2.3",
-    "@angular/platform-browser-dynamic": "18.2.3",
+    "@angular/platform-browser": "19.0.3",
+    "@angular/platform-browser-dynamic": "19.0.3",
     "material-icons": "1.13.12",
     "rxjs": "7.8.1",
     "tslib": "2.7.0",
@@ -23,9 +23,9 @@
   },
   "description": "Shows the ranking of teams for 'MdRzA - Mit dem Rad zur Arbeit'. Users can create a link to share or use himself.",
   "devDependencies": {
-    "@angular-devkit/build-angular": "18.2.3",
-    "@angular/cli": "18.2.3",
-    "@angular/compiler-cli": "18.2.3",
+    "@angular-devkit/build-angular": "19.0.4",
+    "@angular/cli": "19.0.4",
+    "@angular/compiler-cli": "19.0.3",
     "@cypress/schematic": "2.5.2",
     "@types/jasmine": "5.1.4",
     "angular-eslint": "18.3.0",
